+++ apps/web/lib/mcp/pack-registry.ts
@@ -57,6 +57,7 @@ import { wikiOverlayPack } from "./packs/wiki-overlay-pack";
 import { licensingPack } from "./packs/licensing-pack";
 import { marketingOpsPack } from "./packs/marketing-ops-pack";
 import { coworkerCapabilityPack } from "./packs/coworker-capability-pack";
+import { publicWebDesignPack } from "./packs/public-web-design-pack";
 import { sorReadPack } from "./packs/sor-read-pack";
 
 import { grokSigninPack } from "./packs/grok-signin-pack";
@@ -103,6 +104,7 @@ export const TOOL_PACK_REGISTRY = composeToolPacks([
   licensingPack,
   marketingOpsPack,
   coworkerCapabilityPack,
+  publicWebDesignPack,
   grokSigninPack,
   sorReadPack,
 ]);
