+++ scripts/mcp-tool-pack-baseline.json
@@ -7,8 +7,6 @@
   "admin_run_migration",
   "admin_run_seed",
   "admin_view_logs",
-  "analyze_brand_document",
-  "analyze_public_website_branding",
   "analyze_reusability",
   "approve_decomposition",
   "assess_archetype_refinement",
@@ -31,16 +29,12 @@
   "dpf_test_kernel_refuse_probe",
   "drive_browser_task",
   "edit_sandbox_file",
-  "evaluate_page",
   "evaluate_tool",
   "execute_promotion",
-  "extract_brand_design_system",
-  "fetch_public_website",
   "find_related_tests",
   "generate_code",
   "generate_codebase_manifest",
   "generate_custom_archetype",
-  "generate_design_system",
   "get_backlog_item",
   "get_build_dispatch_history",
   "get_build_engine_readiness",
@@ -101,10 +95,8 @@
   "schedule_promotion",
   "schedule_release_bundle",
   "search_code_graph",
-  "search_design_intelligence",
   "search_portfolio_context",
   "search_project_files",
-  "search_public_web",
   "search_sandbox",
   "search_specs_and_plans",
   "search_tool_marketplace",
